jimek/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jimek"
version = "0.1.0"
description = "Advanced process orchestrator with scheduling, multi-channel notifications, and Luigi workflow integration"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "AI Playground" }
]
keywords = [
    "orchestrator",
    "scheduler",
    "workflow",
    "luigi",
    "notifications",
    "automation",
    "pipeline",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Distributed Computing",
]

dependencies = [
    # Core scheduling
    "apscheduler>=3.10.0",

    # Workflow management
    "luigi>=3.5.0",

    # HTTP client for notifications
    "httpx>=0.27.0",

    # Configuration
    "pyyaml>=6.0.0",

    # Data validation (optional but recommended)
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
# Database job stores
db = [
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
]

# Redis job store
redis = [
    "redis>=5.0.0",
]

# All notification dependencies
notifications = [
    "httpx>=0.27.0",
]

# Development dependencies
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-PyYAML>=6.0.0",
]

# All optional dependencies
all = [
    "jimek[db,redis,notifications,dev]",
]

[project.scripts]
jimek = "jimek.cli.main:main"

[project.urls]
Homepage = "https://github.com/ai-playground/jimek"
Documentation = "https://github.com/ai-playground/jimek#readme"
Repository = "https://github.com/ai-playground/jimek"
Issues = "https://github.com/ai-playground/jimek/issues"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["."]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["jimek"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
